Rebrand as routing tree web toolkit
This drops the framework part.  It's mostly semantics, but
framework toolkit makes Roda appear to be designed to build
frameworks as opposed to building applications.  While you
certainly can build frameworks on top of Roda, I think most
users build applications on top of it, so it makes sense to
focus on that.