Make middleware plugin also run the application's middleware
This was an oversight of the initial design.  It's important to
use the Roda application's middleware when running the Roda
application as middleware.  For example, if the Roda application
is using the csrf plugin, when the Roda application was used as
middleware previously, the CSRF plugin wouldn't take effect (since
it is implemented via middleware). This fixes things so that
using the Roda application as middleware uses the same middleware
as when running the Roda application directly.