Make :header matcher prefixes the env key with HTTP_ if application :header_matcher_prefix option is set
It is not backwards compatibility to automatically prefix the value,
as users may already be prefixing with http- manually, as shown in
the specs.

It is possible to check both the prefixed and unprefixed values,
but I'm not sure that will not add a security risk in some code
where there is a value for the fallback key but other code is
checking the primary key.

It seems safest to handle this via an application option.  The
behavior can switch to assuming the option is true starting in
Roda 3.