Change example code in match_affix documentation (Fixes #83)
The example previously shown results in a routing tree that doesn't
really work.  It's OK to consume an slash from a segment if you are
changing the prefix to not include a slash, but doesn't make sense
to do so unless the slash is at the end of the path if you are
expecting matches to start with a slash.