Add nocov markers around a rescue block in the public plugin
I guess File.file? and File.readable? can raise SystemCallError
in extreme cases, so the code should probably handle that, but
I'm not sure how the tests are supposed to check for that other
than stubbing one of the methods.  Since there are unlikely to be
bugs in a literal false, I think marking this nocov is safe.