Fix assets_paths method in assets plugin when subresource integrity is used
When assets_paths was split from assets, the logic that handled
subresource integrity was moved as well, even though that's not
part of the path.  Basically, subresource integrity was kind of
hacked in before, but it wasn't a bug until assets_paths was added.

To fix things, add a private _compiled_assets_hash method, which
returns either the hash itself or the unique key that used in the
path, and have both assets and assets_path call it if necessary.