Add support for :unsupported_block_result => :raise option, for raising on unsupported route/match block return values
This can make it easier to find potential issues in your code,
though it may require more explicit code.  By default, if you
use a block return type that is not supported, it will just be
ignored, which can be frustrating if you don't realize what
the problem is.

I'm considering making this the default behavior in Roda 3.