Make error_email method in error_email plugin support non-exception arguments
Sometimes you may want to send an email when there is an error
condition but an exception hasn't been raised.  For example, if
there is a certain condition in your application for which there
should be a matching object for a hash key, but no such object
exists.  While you could raise an exception to handle it, you
may just want to send the error email without raising an
exception.