Allow middleware plugin to accept a block that will be used to configure the application when used as middleware
This should allow configuration of any part of the middleware, since
the user has complete control via the block.  You can even choose
to accept blocks when the middleware is loaded into an application,
and customize how that block is handled.

Since configurable middleware would probably change the Roda app,
use a subclass of the Roda app for any configurable middleware.
This allows the Roda app to be used as middleware with multiple
separate configurations.

Disallow accepting arguments when loading as middleware unless
the middleware plugin was initialized with a block.