Handle extensions containing dots in type_routing
The type_routing plugin previously wouldn't handle extensions containing
the period character. For example, in this app:

    class ExampleApp < Roda
      plugin :type_routing, types: {
        :gz => 'application/octet-stream',
        :'tar.gz' => 'application/octet-stream',
      }
    end

URLs ending in ".tar.gz" would always be detected as the "gz" type,
never the "tar.gz" type.

The commit changes the regex that matches the URL. The group that
captures the "basename" was greedy (by default), meaning that it would capture
`["whatever.tar", "gz"]`. Adding the '?' changes the capture group to be
lazy, so it will now capture `["whatever", "tar.gz"]` instead.