Add class_matchers plugin for matching other classes (in addition to String/Integer), with user specified regexps and type conversion
This generalizes the class matching code to dispatch by name,
which makes extension easier.  To allow consume to handle type
conversion, have it yield to a block if a block is given, and
use the block return value.  This avoids duplication for handling
the Integer class matcher.

The class_matcher plugin is similar to symbol_matchers, except
that it supports type conversion and does not support integration
with the params_capturing plugin.