Simplfy locals handling in render_locals plugin
If merging locals, the plugin locals are already merged into the layout
locals in configure, so no need to merge them again.  This does slightly
change behavior if the :layout_opts=>:merge_locals method option is
given, but I would say the expected behavior in that case is to only
merge the method locals and not the plugin locals, which is what
this code still does.

This also removes the if block for layout locals, since the plugin
will always have layout locals defined, even if empty.