Allow capturing errors during convert! and convert_each! and raising at the end
This requires significant internal API changes so that exceptions are
recorded instead of immediately raised inside convert! and
convert_each! blocks.  The default value is now passed all the way
down to process, and a CHECK_NIL sentinal value is used to immediately
raise/record the exception there.

Error#all_errors and Error#param_names have been added to get all
exceptions that would have been raised inside the convert! or
convert_each! block and all the parameter names for those
exceptions.

Params#[] would not work well inside convert blocks, since it
raises an immediate error, and can't just record the error and
return nil, since it is generally chained to another method.
To work around that issue, convert! now takes an array of subkeys
for parameter traversal, and yields the resulting value to the
block.  If there is no value, it just records the error and
returns nil (or raises the error if not called inside another
convert! block).

Add a protected method named subkey to handle parameter traversal,
and have both dig and convert use it.

This also removes some now unnecessary methods:

* Error.create
* Params#nesting
* Params#obj
* Params#start_capture
* Params#check_value

Calling #dig with :[] no longer works because #dig now expects
to be called with a conversion method.  #dig's method calling
will probably be made more strict in the near future.