In typecast_params, add Error#reason for a symbol containing the reason for the error
This makes it simpler to determine why an error occurred, and could be
used to provide nicer error messages to the user.

Have handle_error take the reason as an argument, and convert callers
to use it.

Change Error#all_errors so that it always returns an array where the
first entry is the receiver. This makes it much simpler to use.

Rename Error.set_keys to Error.create, now that it handles both the
keys and the reasons.