plugins/head: call .close on response body
Calling the `close' method on the response body may be used for
resource management or logging purposes; not calling it can
cause resource exhaustion in some applications.

We delay calling `close' until the Rack server deems it
necessary to, so introduce a CloseLater wrapper.  This avoids
potential race conditions in case the response headers require a
locked resource (which `close' would release) and improves
timing accuracy for Rack::CommonLogger in rack 1.4.0+

rack 1.4.0+ has a Rack::BodyProxy class which is used by the
Rack::Head middleware since rack 1.6.0 to handle this.  I am not
relying on Rack::BodyProxy at the moment since our rack
dependency is unversioned.