Add early_hints plugin for sending 103 Early Hint responses, and :early_hints assets plugin option
When used with puma 3.11+ with the --early_hints flag, these allow
you to have 103 Early Hints responses before the 200 response for
all assets added by calls to the assets method.

Currently, the :early_hints assets plugin option is an all or
nothing setting, but it's fairly easy to add a per-call override
if needed.