Add :wrap option to json_parser plugin, for whether/how to wrap the uploaded JSON object (Fixes #142)
This supports values of either :always to always wrap or :unless_hash
to only wrap values that are not hashes.  I think if you are going to
wrap, you should use :always so that you get consistent behavior.
Only wrapping non-hash uploaded is inconsistent and probably a bad
idea, though it can make the common case easier.