Add :handle_result option to middleware plugin for modifying rack result before returning it
Previously, the middleware plugin could be used to return results
if the roda app handled the request, but if the request was passed
to the  next middleware, the result from that was returned without
modification.  This adds the ability to modify the result, just
as a typical rack middleware can.