Add sessions plugin for more secure (encrypted+signed) sessions
This adds a session plugin, now recommended instead of
Rack::Session::Cookie whenever sessions are required.  It uses
sessions encrypted with the AES-256-GCM cipher, and signed with
HMAC-SHA-256.  Both signing and encryption are used so that
attacks on the encryption cannot be performed without first
successfully forging a valid signature.

In addition to encryption and a stronger default signature
algorithm compared to Rack::Session::Cookie, the sessions
plugin has the following benefits:

* Built in session expiration enabeld by default (default:
  30 days since session creation, 7 days since last update)

* Padding by default to minimize information leakage due to
  differing session data sizes (session data padded to
  multiples of 32 bytes by default)

* Automatic compression of large sessions via deflate (by
  default if session over 128 bytes)

* JSON is used for serialization by default instead of
  Marshal, preventing remote code execution vulnerabilities if
  the secrets are disclosed. Note that this means that symbols
  do not round trip in the session, which will probably be the
  largest barrier to adoption.

* A plain hash is used for the session, instead of a hash-like
  object. One consequence of this is that keys in the session
  are not automatically converted to strings.
  Rack::Session::Cookie converts session keys to symbols for
  keys at the top level, but not for keys in subhashes.

* In general sessions are smaller, despite requiring an
  additional 28 bytes for the cipher initialization vector
  and auth tag. The main reason for this is that the sessions
  do not set a session id, since one is not needed for cookie
  sessions.

The sessions plugin ties into the Roda#session method instead
of being a rack middleware.  This makes it about twice as
fast as Rack::Session::Cookie if the session is not accessed.
If the session is accessed, the sessions plugin is roughly
as fast as Rack::Session::Cookie (within 10%), even though
it uses a stronger HMAC and has to encrypt and decrypt the
session.

Because the sessions plugin is not a middleware, it does not
offer session support to other middleware, only to the app
itself.  If you would like to use the same approach as the
sessions plugin uses but would like support for middleware to
access the sessions, a roda/session_middleware file has been
added, which adds RodaSessionMiddleware, which is a middleware
that can be used by any other Rack app for session support, and
which uses a SessionHash class similar to the one used by
Rack::Session::Cookie.