Better handle upgrades from :_flash key
Now that the '_flash' session key is deleted if the flash is empty,
make sure to delete the :_flash key so and old flash won't be used,
and also to make sessions smaller by removing unneeded data.

This is only going to be necessary on session implementations that
don't convert symbol keys to string keys and support symbol keys.
So this change should not affect usage with the sessions plugin,
session_middleware, or Rack::Session::Cookie.