Use a single :secret option in sessions plugin
Instead of requiring :hmac_secret and :cipher_secret to be
specified separately, just require a longer :secret option, and
use the first 32 bytes as the :cipher_secret, and the rest as
the :hmac_secret.

Similarly, support :old_secret and split it to create
:old_cipher_secret and :old_hmac_secret.

One added benefit is that since we are munging the secrets, they
can also be frozen and forced to binary encoding.