Refactor how internal hooks are implemented
Instead of having plugins override Roda#call and make behavior
dependent on plugin load order, this adds internal _before_hook
and _after_hook plugins, and has the related plugins use them.

The _(before|after)_hook plugins scan for private instance
methods in the class named _roda_before_* and _roda_after_*,
and call those methods in order during before and after
processing.

There is now a fixed order for before processing when related
plugins are used:

* hooks
* heartbeat
* static_routing

and a fixed order for after processing:

* class_level_routing
* status_handler
* head
* flash
* session
* hooks

This allows users to not have to worry as much about plugin
load order, and simplifies some of the internals.

Some of the before processing used to be called before the
catch block in Roda#call, and some inside the catch block.
Now, all before hooks are run inside the catch block.
To implement this, a rack_app_route_block private class
method has been added, which plugins can override to modify
the default route block to add their own handling. This
reduces block creation during routing.

To allow route processing without rerunning before and after
internal hooks, a Roda#_call private method has been added.
Plugins that used to override Roda#call and call super with
a different block in some conditions have been switched to
use Roda#_call.

The error_handler plugin now loads the _after_hook plugin,
and it is now the only plugin other than the _after_hook
plugin to override Roda#call.  This is because the
error_handler has to use a rescue around the call to super,
it cannot operate as an after hook.  However, to keep
backwards compatibility, after the error_handler is run, it
calls back into normal after processing, with the exception
that any errors at that stage are logged but otherwise
ignored instead of being raised.

The _roda_after_* methods are now called even if the rack response
is nil because an exception was raised.  This is kind of unfortunate,
but necessary for hooks plugin backwards compatibility.  The affected
code has been modified to check for a present rack response.

In addition to fixing plugin load order issues, this
approach can also reduce the call stack in cases where
multiple plugins listed above are in use, which may help
shorten garbage collection time.