Disable default compression of sessions over 128 bytes in the sessions plugin
If the application puts user-derived data in the session, and is
storing sensitive data in the session, then compression is a bad
idea from a security perspective as it allows for compression
ratio attacks. Compression ratio attacks are mitigated by the
default padding post-compression of random data, and the average
application does not store both sensitive data and user-derived
data in the session, but it's best to eliminate the possibility.

Applications can still enable compression by setting :gzip_over
to an integer.

This change doesn't affect backwards compatibility, as compressed
sessions will still be decrypted, and there is no security risk
there.