Do not persist convert!/convert_each! :symbolize setting in the typecast params plugin
Previously, this setting was sticky, because capturing works by
recursing into subhashes after the capture process is complete.
Save the then-current nested_params hashes before resetting the
symbolize setting so that when the recursion happens, the
correct values are used.

This can break backwards compatibility if you were relying on
the previous behavior, but the previous behavior wasn't
intended and is more likely to cause bugs than be intended.