Allow set_layout_opts in view_options plugin to override layout if render plugin :layout option is given
Previously, this didn't work because the :layout option in the
render plugin would override the layout options set by
set_layout_opts.  The layout template is already available via
:layout_opts :template, there is no reason it also needs to be
set via :layout as well.

It is possible that some existing plugins could be looking at
:layout render plugin option instead of :layout_opts :template
to get the default layout.  This change could break that case,
but should not have a negative effect beyond introspection of
the plugin options.