Do not have RodaResponse#initialize set status to nil
This is better for performance as the value of an uninitialized
instance variable is nil by default.

This can result in backwords incompatibility if code previously
expected the status to be set back to nil.  However, initialize is
a private method, so this is not considered a breakage of the
public API.