Add Roda.define_roda_method for defining instance methods instead of using instance_exec
instance_exec is slow as it creates a singleton class for every
object you call it on.  The only time you really need to use
instance_exec is if you don't know the class of the instance
you want to use to execute the block in advance, or you really
will be executing the block on only a specific instance.