Improve performance by using method instead of instance_exec for main route block
This introduces a new internal dispatch API for Roda.  Instead of
using Roda#call with the route block (which basically requires
instance_exec), define a method for the route block, and just call
that as a normal method.  The following instance methods are
added:

  * _roda_handle_main_route: Entry point for normal request dispatch.
  * _roda_handle_route: Yields to the routing block, catching any
    halts inside the block, treating the block as a routing block.
  * _roda_main_route: Roda.route defines this method using the
    block provided, it accepts the request as an argument.
  * _roda_run_main_route: Calls _roda_main_route with the request.

It is now a warning to call Roda.route without a block, and a
warning to have a module included in the class which overrides
Roda#call (the old dispatch API) but does not override
Roda#_roda_handle_main_route or Roda#_roda_run_main_route. Both
of these will be errors starting in Roda 4.

Many of the plugins that ship with Roda still use the old dispatch
API, these will be fixed shortly.