Improve performance of internal before and after hooks, integrate internal after hooks with error_handler plugin
If there is only a single internal before or after hook defined, use
an alias instead of defining a new method to save a method call.

Integrate the internal after hook plugin into the error_handler
plugin.  This is faster if the error_handler plugin is used, but
it is slower if the internal after hook plugin is used without
the error_handler plugin.  I expect it is much more common to
use the error_handler plugin than it is to use the internal
after_hook plugin without the error_handler plugin.  This
approach is also a bit simpler.

This moves the internals for the internal before hook plugin
back into the plugin file instead of being implemented in
core, since it now can be done more efficiently that way.

This updates both the internal before hook plugin and
the error handler plugin to use the new dispatch API for
better performance.