Improve performance of route_csrf plugin by using a method for default failure action
If a :csrf_failure option value is passed as a Proc to the check_csrf!
method, continue to instance_exec it, but warn in this case.  Support
for this was not intentional, it was a side effect of the
implementation.  Now that a method is used for specifying the failure
behavior, there is no reason to continue to support it.

While here, have the plugin block integrate with the route_block_args
plugin, since it operates as a route block.