Improve performance of the static_routing plugin by using methods instead of instance_exec
This also removes the r.static_route method.  The API it used
would not be compatible with the new use of methods.  This
method was not designed to be called by external code.

It is now an error to call one of the static_* class methods
without a block.  Previously, the error would not be hit
until runtime.