Add support for per-cookie cipher secrets in the sessions plugin, and enable them by default
The sessions plugin previously used the same cipher secret/key
for all encryption.  That is not normally a problem as long as
the initialization vector used in each encryption is truly random,
as the odds of an initialization vector being used more than once
would be 1 in 2**64.  However, if there is a weakness in the random
number generator, and the initialization vector is used more than
once, that can result in exposure of the cipher secret/key.

This new approach avoids the problem.  It adds 32-bytes of random
data to the cookie, and generates the cipher secret/key for the
encrypted data for that cookie using HMAC-SHA-256 with the default
cipher secret on the random data.  In the case where the random
number generator is broken, and an initialization vector is reused,
this would not leak the default cipher secret/key, it would only leak
the cipher secret/key for that particular cookie, which would not
be usable to decrypt any other cookies unless the random data used
as the basis for the cipher secret for those cookies was the same.

This change is not without costs. In terms of storage overhead,
the additional 32 bytes can make the cookie up to 50% larger if
the cookie is almost empty, and probably about 15% larger in the
average case where you are storing a login ID and a CSRF token.

In terms of processing overhead, this adds about 10% to the
serialization and deserialization of cookies, so about 20% total
if you are modifying a cookie during a request.  However, in most
applications, time spent serializing and deserializing session
cookies will be minimal compared to other processing.

As this enables the per-cookie cipher secrets by default, it can
result in issues if you are doing a rolling restart of applications,
where some processes only support version 0 sessions (no per-cookie
secrets) and other processes support version 0 and 1 sessions
(per-cookie secrets).  If a process that supports version 0 and 1
sessions receives a request and returns a response with a version 1
session, and the next request is received by a process that only
supports version 0 sessions, that process will not be able to
decode the session.  To avoid that, when upgrading, temporarily
set :per_cookie_cipher_secret to false, then restart all processes.
After all processes have restarted, set :per_cookie_cipher_secret
to true, and restart all processes again.