Add hash_routes plugin for O(1) route dispatching at any level in the routing tree
This takes the performance approach of the static_routing plugin
with the flexibility of the multi_route plugin.  It also offers
a DSL that looks quite similar to Roda's standard DSL, to make
the plugin easier to use.

The views method in this plugin can be a faster replacement for
the multi_views plugin.