Delay building rack app until Roda.app is called
This delays building the rack app until Roda.app is called. This
is similar to the previous behavior of the delay_build plugin,
except that now building happens automatically without having
to call a method.

This can fix O(n^2) issues when building applications with a
lot of middleware.

To make this work, anything that could change the rack app unsets
the existing rack app, such as adding a plugin, modifying
middleware, or changing the route block.

One consequence of this change is that calling Roda.route is
optional, so you are no longer forced to call it with an empty
block if you are doing handling outside the route block, such
as using the static_routing plugin or before hook to handle all
requests.