Improve string matching by 10-20%
This changes _match_string to only perform a single allocation,
and only in the successful case of matching a non-final segment.
Most ruby operations are equality comparisons on integers.

The previous code allocated a string for rp.start_with? and
a string for rp[last], and checked for the segment end using
a string operation instead of an integer operation.