Reorder case clause in RodaRequest#match
Deprioritize Symbol, as it is recommended to use class matchers
instead of Symbols.

Prioritize true, as that is often used with verb methods to
indicate the path has already been fully matched.

Prioritize false and nil over proc, as it is better to avoid
using procs unless they are passed from elsewhere, and then
are not that useful as they would have no access to request
specific information.

Prioritize Array over Hash, as array usage for matching
any elements seems more common that hash usage.

If anyone really cares about a different order, we can add
a plugin to specify the case order.