Improve performance for handling blocks with arity 0 where expected arity is 1 by avoiding instance_exec
Instead of using instance_exec, just define a method for the block,
which will also have 0 arity, and define a regular method that will
call that method either directly if possible, or via send if not.

This is around 2.75x faster if a singleton class has already been
created for the instance, and over 8x faster if a singleton
class has not been created (as instance_exec will create one in
that case).

It's still faster by about 20% to define the block with the
correct arity of 1, and so Roda will continue to warn about the
arity mismatch if :check_arity is set to :warn.