Set Content-Length header to 0 for empty 205 responses
RFC 7232 requires one of three actions for 205 responses:

1) Set Content-Length to 0
2) Use Transfer-Encoding chunked with a single empty chunk
3) Close the connection after sending the headers

We can't control 3), and 1) seems easier than 2).

This changes Roda's behavior for 205 responses to set the
Content-Length to 0.  It still deletes the Content-Type
header.  By default, this is only done for empty responses,
but if the drop_body plugin is used, it will do so for all
responses.

Inspired by similar changes in rack and Sinatra.