Cache compiled templates in development mode, until the template files are modified
This can significantly improve performance in development mode,
especially with complex templates or templates that are slow to
compile.

Implement this using a wrapper object that checks if the file
still exists and checks the mtime of the file if so.  If the
file exists and the mtime is different, the wrapper uses a new
template object.