Improve render performance up to 4x in the default case by calling compiled template methods directly
This takes the UnboundMethods that Tilt prepares for templates,
and defines real methods using them in a module included in the
Roda class (so it works with frozen Roda apps).

When render/view is called with a single Symbol or String argument,
lookup that argument in the template method cache.  If it exists,
call the method using send.  If it doesn't exist, but a single
Symbol or String argument is given, flag to update the template
method cache with the compiled template method.

This optimizes the common case of calling render/view with a
single argument.  It can be 2.5x faster to render and 4x
faster to view for simple templates in terms of time taken for
the entire request processing.

This depends on Ruby 2.3+ and Tilt 1.2+ in order to work correctly,
so only use the compilation in those cases.  As this calls a private
Tilt method, check that the method has the expected arity before
enabling this optimization.

Most plugins that use render don't benefit directly from this
optimization, but I expect many apps benefit significantly from
it.  The render_locals plugin turns the optimization off completely,
because it can inject locals even when none are passed directly when
calling render/view, and the optimization does not work in that case.

If using the view_options plugin, the set_view_subdir and
append_view_subdir can work with the optimization, but
set_view_options and set_layout_options disables the optimization.

Some less commonly used tilt template engines do not support
compilation.  For those template engines, if compiling the
template raises a NotImplementedError, rescue it and mark
to not try compiling the template in the future.

To avoid problems when subclassing, make sure to define methods
that include the class object id in the template method cache key.