Improve render/view performance up to 2x in development mode in the default case by calling compiled template methods directly
This extends the earlier approach for using compiled methods to
development mode.  Instead of defining a single compiled method,
per template, two compiled methods are defined.  The main compiled
method will check for modifications to the template file, and if
so, it will generate a new compiled template method with the
updated template file, then call that method.

Speed up the check for template file modification, using a single
File.mtime call with rescue instead of File.file? and File.mtime.
This is faster and avoids a race condition, and in most cases the
file will still exist.