Improve render_each performance by calling compiled template methods directly
This improves render_each performance for a single simple render by
about 2x, and render_each performance for an enumerable with 100 objects
by 3x (cache: false) to 9x (cache: true).

Currently, this optimization is only done if no options are passed
render_each, just as is done for render/view.