Improve render performance with :locals option up to 75% by calling compiled template methods directly
This has the render plugin recognize a single :locals option and
also optimize that case (though not as well as the no option case).
Performance for simple templates can be up to 75% better using this
approach.

Refactor the render_each support for compiled templates to use
this new support in the render plugin, dramatically simplifying
the code.