Support change in tilt 2.0.10 private API to continue to support compiled templates, with up to 33% performance improvement
Tilt 2.0.10 was just released with patches added by me that fix
warnings on Ruby 2.7 and improve performance of compiled
template methods.  These patches change the private API in tilt
to pass in the class to use for constant resolution inside the
compiled method.

Adjust the render plugin to support this change in private API,
while continuing to support the previous private API that is
supported back to Tilt 1.2.  Add a
Render.tilt_template_compiled_method method to handle this, and
define separate versions depending on the private API supported.

Update the precompiled_templates specs to handle the change to
Tilt's internals.