Add asynchronous streaming to the streaming plugin
Generating the response body can be time consuming and add overhead to
the total download duration. ActionController::Live in Rails performs
streaming in a background thread, which mitigates this concern. So we
add this functionality to the streaming plugin, making it opt-in via the
:async option.

The response body generation starts already on initializing the stream
object in order to minimize overhead. By default we use a sized queue
for backpressure, to keep memory growth under control. We use the
default size of 10, since that's what ActionController::Live uses.

However, we allow the developer to specify their own queue object. This
is most useful with Falcon, with which it's desirable to use an
Async::Queue instead of a regular Queue, as then the streaming is fully
non-blocking.