Add multibyte_string_matcher plugin for matching multibyte characters
multibyte characters are not commonly used in paths.  For one,
they have to be URL escaped by browsers.  Roda previously handled
multibyte characters in request paths correctly, but an
optimization added in 3.19 broke the handling of multibyte
characters.

Introduce a plugin for this support, so that only people that need
to handle multibyte characters in request paths pay the cost for
the support.