Make sure all requests in specs are made with PATH_INFO set
This is not a requirement for testing, but it allows testing
the effect of plugins with all specs loaded, useful for determining
if loading a plugin can break unrelated code.  Some plugins require
that PATH_INFO be provided to work correctly (the rack spec
requires PATH_INFO, so plugins requiring it is reasonable).