Add each_partial to the partials plugin for rendering a partial for each element in an enumerable
If you wanted to use partials and render_each previously, you would
have to manually use an underscore in the template filename and
specify the :local option when calling render_each.  Like the
partial method, this is just a simpler way if you want to use the
convention of partial filenames being prefixed with an underscore.