Make :dependencies option in assets plugin work correctly with render plugin template caching (Fixes #191)
First, add another entry to assets_opts that contains the expanded path
of the :dependencies option.  This is needed because the assets plugin
indexes into the hash with an expanded path.

Refactor TemplateMtimeWrapper slightly in the render plugin. Precompute
the dependencies array if there are dependencies to save work at runtime.
Make template_last_modified no longer take a path, as that was just done
as an optimization.  This breaks backwards compatibility, but this is
not a class that should be accessed directly.

The previous commit had :dependencies be a render plugin option, but that
doesn't make sense unless all templates you are rendering depend on the
same file, which seems unlikely.  Switch this to a render method option,
since that is what the assets plugin needs to specify on a per-call basis.