Add precompile_views and freeze_template_caches! to the precompile_templates plugin
precompile_views precompiles the optimized render methods, so they
can also be shared by child processes.

freeze_template_caches! ensures that all templates are precompiled
(raising an error if trying to compile a template afterward).  It
also speeds up access to the caches by avoiding the use of a mutex
when accessing them.