Improve view performance with :content option up to 3x by calling compiled template methods directly
This is implemented so it has no impact if view is called with
multiple arguments.  There is a less than 1% impact when calling
view with a single template name argument.  I considered adding a
separate content_view method for this, to avoid this small decrease,
but determined it wasn't a good tradeoff.

This adds a private _optimized_view_content method that returns nil
when the optimized support is not available (old versions of Ruby
or Tilt), or when the optimized support should not be used as it
can result in incorrect results (such as when the render_locals
plugin is used).

This also unsets the :_roda_layout key in the template method
cache if it is already set, so that changing the layout options
by reloading the plugin will not keep the old compiled method.