Make the time of host_authorization check configurable
This makes the host authorization check an instance method.  This
is similar to the route_csrf plugin, where the user has control
where the check is done.  Doing this makes the plugin more flexible.
This can result in faster code if you can delay the check until after
routes that may not need it (such as static file serving routes). It
also allows for checking host authorization in before hooks, which
previously was not possible.